Consolidate pull request utilities
